Basic structure for peer-to-peer connections
Created the EkeUser class used to handle a single EkeUser.
It contains two threads: one for the server part and one for the client part.
The listening socket is working and available for others, now I'm composing how
the users will cooperate with each other.
Next step: menù structure and basic communication.
◴_◶