Menù and names
Added the menus and the execution flow of the program.
Having some issues assigning the name from console and not from stdin: to be
checked.
Tried sending some data from one client to the other and seem to work, but I
still need to figure out how correctly it has to be done.
(҂⌣̀_⌣́)