Done!
The EKE authentication is correctly working!
The problem with the encryption/decryption was correlated to the encryption key
and iv: it's wrong on Ruby's documentation, but I corrected it and now it's
working.
All packets have been implemented and are correctly exchanged between the two
peers.
Up to five instances of the program could run at the same time.
Next step: refactoring and compute G, the Zp* generator.
(｡し_し｡)